all update

+++ DAL/Model/ReportItem.cs
@@ -4,6 +4,6 @@ public class ReportItem
 {
     public string IngredientName { get; set; }
     public string UnitName { get; set; }
-    public decimal TotalQuantity { get; set; }
-    public decimal TotalCost { get; set; }
+    public double TotalQuantity { get; set; }
+    public double TotalCost { get; set; }
 }
